test(testing): Shell и WebViewScreen tests вынесены в отдельные suite files
shell-level app state и navigation tests, а также web view navigation и state tests выделены из RSSReaderTests.swift в самостоятельные feature-oriented suites
