+++ RSSReaderTests/RSSReaderTests.swift
@@ -7,213 +7,6 @@ import Testing
 @MainActor
 struct RSSReaderTests {
     @Test
-    func readingShellNavigationStateBuildsDetailDestinationsForNoneAndArticleRoutes() {
-        let articleID = UUID()
-
-        #expect(
-            ReadingShellDetailNavigationState.detailDestination(
-                route: .none,
-                selectedArticleID: nil
-            ) == .none
-        )
-        #expect(
-            ReadingShellDetailNavigationState.detailDestination(
-                route: .none,
-                selectedArticleID: articleID
-            ) == .article(articleID)
-        )
-        #expect(
-            ReadingShellDetailNavigationState.detailDestination(
-                route: .article(articleID),
-                selectedArticleID: nil
-            ) == .article(articleID)
-        )
-    }
-
-    @Test
-    func readingShellNavigationStateBuildsWebViewDestinationForWebViewRoute() {
-        let route = ArticleWebViewRoute(
-            articleID: UUID(),
-            url: URL(string: "https://example.com/web-shell-destination")!
-        )
-
-        #expect(
-            ReadingShellDetailNavigationState.detailDestination(
-                route: .webView(route),
-                selectedArticleID: route.articleID
-            ) == .webView(route)
-        )
-    }
-
-    @Test
-    func webViewScreenNavigationStateClosesOnLeftEdgeHorizontalDrag() {
-        #expect(
-            WebViewScreenNavigationState.shouldCloseOnDrag(
-                startLocationX: 12,
-                translation: CGSize(width: 96, height: 10)
-            )
-        )
-    }
-
-    @Test
-    func webViewScreenNavigationStateIgnoresDragsAwayFromLeftEdge() {
-        #expect(
-            WebViewScreenNavigationState.shouldCloseOnDrag(
-                startLocationX: 48,
-                translation: CGSize(width: 96, height: 8)
-            ) == false
-        )
-    }
-
-    @Test
-    func webViewScreenNavigationStateIgnoresMostlyVerticalDrags() {
-        #expect(
-            WebViewScreenNavigationState.shouldCloseOnDrag(
-                startLocationX: 10,
-                translation: CGSize(width: 96, height: 64)
-            ) == false
-        )
-    }
-
-    @Test
-    func webViewScreenStateBuildsInitialDerivedStateFromRoute() {
-        let route = ArticleWebViewRoute(
-            articleID: UUID(),
-            url: URL(string: "https://example.com/articles/webview-state")!
-        )
-        let state = WebViewScreenState(route: route)
-        let viewState = state.derivedViewState()
-
-        #expect(viewState.initialURL == route.url)
-        #expect(viewState.navigationTitle == "example.com")
-        #expect(viewState.primaryLoadingState?.title == "Loading Page")
-        #expect(viewState.placeholder == nil)
-        #expect(viewState.loadingProgress == 0)
-        #expect(viewState.reloadRevision == 0)
-        #expect(viewState.showsWebViewContent)
-        #expect(viewState.showsShareAction == false)
-        #expect(viewState.showsBottomActions == false)
-        #expect(viewState.toolbar.shareURL == route.url)
-        #expect(viewState.toolbar.isShareEnabled)
-        #expect(viewState.bottomActions.isRefreshEnabled)
-        #expect(viewState.bottomActions.openExternalBrowserURL == route.url)
-        #expect(viewState.bottomActions.isOpenExternalBrowserEnabled)
-    }
-
-    @Test
-    func webViewScreenStateTracksTitleLoadingProgressAndFailureIndependentlyFromView() {
-        let route = ArticleWebViewRoute(
-            articleID: UUID(),
-            url: URL(string: "https://example.com/articles/webview-state-progress")!
-        )
-        var state = WebViewScreenState(route: route)
-
-        state.applyNavigationStart()
-        state.applyLoadingProgress(0.42)
-        state.applyPageTitle("Loaded Article")
-
-        var viewState = state.derivedViewState()
-        #expect(viewState.primaryLoadingState?.title == "Loading Page")
-        #expect(viewState.placeholder == nil)
-        #expect(viewState.loadingProgress == 0.42)
-        #expect(viewState.navigationTitle == "Loaded Article")
-        #expect(viewState.showsShareAction == false)
-        #expect(viewState.showsBottomActions == false)
-
-        state.applyNavigationFinished()
-        viewState = state.derivedViewState()
-        #expect(viewState.primaryLoadingState == nil)
-        #expect(viewState.placeholder == nil)
-        #expect(viewState.loadingProgress == 1)
-        #expect(viewState.showsShareAction)
-        #expect(viewState.showsBottomActions)
-
-        state.applyNavigationFailure("The page could not be loaded.")
-        viewState = state.derivedViewState()
-        #expect(viewState.primaryLoadingState == nil)
-        #expect(viewState.placeholder?.title == "Failed to Load Page")
-        #expect(viewState.placeholder?.description == "The page could not be loaded.")
-        #expect(viewState.showsWebViewContent == false)
-        #expect(viewState.showsShareAction == false)
-        #expect(viewState.showsBottomActions == false)
-    }
-
-    @Test
-    func webViewScreenStateSwitchesShareAndBrowserActionsToCurrentPageURL() {
-        let route = ArticleWebViewRoute(
-            articleID: UUID(),
-            url: URL(string: "https://example.com/articles/original")!
-        )
-        var state = WebViewScreenState(route: route)
-
-        state.applyCurrentPageURL(URL(string: "https://example.com/articles/redirected")!)
-
-        var viewState = state.derivedViewState()
-        #expect(viewState.toolbar.shareURL?.absoluteString == "https://example.com/articles/redirected")
-        #expect(
-            viewState.bottomActions.openExternalBrowserURL?.absoluteString
-                == "https://example.com/articles/redirected"
-        )
-
-        state.applyCurrentPageURL(URL(string: "mailto:hello@example.com")!)
-
-        viewState = state.derivedViewState()
-        #expect(viewState.toolbar.shareURL == nil)
-        #expect(viewState.toolbar.isShareEnabled == false)
-        #expect(viewState.bottomActions.openExternalBrowserURL == nil)
-        #expect(viewState.bottomActions.isOpenExternalBrowserEnabled == false)
-    }
-
-    @Test
-    func webViewScreenStateIncrementsReloadRevisionOnlyForSupportedURLs() {
-        let supportedRoute = ArticleWebViewRoute(
-            articleID: UUID(),
-            url: URL(string: "https://example.com/articles/webview-reload")!
-        )
-        var supportedState = WebViewScreenState(route: supportedRoute)
-
-        supportedState.requestReload()
-
-        #expect(supportedState.derivedViewState().reloadRevision == 1)
-
-        let unsupportedRoute = ArticleWebViewRoute(
-            articleID: UUID(),
-            url: URL(string: "mailto:hello@example.com")!
-        )
-        var unsupportedState = WebViewScreenState(route: unsupportedRoute)
-
-        unsupportedState.requestReload()
-
-        #expect(unsupportedState.derivedViewState().reloadRevision == 0)
-    }
-
-    @Test
-    func webViewScreenStateStartsInFailurePhaseForUnsupportedInitialURL() {
-        let route = ArticleWebViewRoute(
-            articleID: UUID(),
-            url: URL(string: "mailto:hello@example.com")!
-        )
-        let state = WebViewScreenState(route: route)
-        let viewState = state.derivedViewState()
-
-        #expect(viewState.primaryLoadingState == nil)
-        #expect(viewState.placeholder?.title == "Failed to Load Page")
-        #expect(
-            viewState.placeholder?.description
-                == "This article link can't be opened in the in-app browser."
-        )
-        #expect(viewState.navigationTitle == "Article")
-        #expect(viewState.showsWebViewContent == false)
-        #expect(viewState.showsShareAction == false)
-        #expect(viewState.showsBottomActions == false)
-        #expect(viewState.toolbar.shareURL == nil)
-        #expect(viewState.toolbar.isShareEnabled == false)
-        #expect(viewState.bottomActions.isRefreshEnabled == false)
-        #expect(viewState.bottomActions.openExternalBrowserURL == nil)
-        #expect(viewState.bottomActions.isOpenExternalBrowserEnabled == false)
-    }
-
-    @Test
     func articlesScreenStateStartsWithoutSelectionPlaceholder() {
         let state = ArticlesScreenState()
 
@@ -1011,78 +804,6 @@ struct RSSReaderTests {
                     .month(.wide)
                     .year()
             )
-        )
-    }
-
-    @Test
-    func readingShellCompactNavigationStateSelectsPreferredCompactColumnForCurrentContext() {
-        #expect(
-            ReadingShellCompactNavigationState.preferredCompactColumn(
-                sourceSelection: nil,
-                articleSelection: nil
-            ) == .sidebar
-        )
-        #expect(
-            ReadingShellCompactNavigationState.preferredCompactColumn(
-                sourceSelection: .unread,
-                articleSelection: nil
-            ) == .content
-        )
-        #expect(
-            ReadingShellCompactNavigationState.preferredCompactColumn(
-                sourceSelection: .feed(UUID()),
-                articleSelection: UUID()
-            ) == .detail
-        )
-    }
-
-    @Test
-    func readingShellCompactNavigationStateShowsArticlesBackButtonOnlyInCompactSourceContext() {
-        #expect(
-            ReadingShellCompactNavigationState.showsArticlesBackButton(
-                horizontalSizeClass: .compact,
-                sourceSelection: .starred
-            )
-        )
-        #expect(
-            ReadingShellCompactNavigationState.showsArticlesBackButton(
-                horizontalSizeClass: .regular,
-                sourceSelection: .starred
-            ) == false
-        )
-        #expect(
-            ReadingShellCompactNavigationState.showsArticlesBackButton(
-                horizontalSizeClass: .compact,
-                sourceSelection: nil
-            ) == false
-        )
-    }
-
-    @Test
-    func readingShellCompactNavigationStateRecognizesLeadingEdgeBackSwipeToSources() {
-        #expect(
-            ReadingShellCompactNavigationState.shouldNavigateBackToSourcesOnDrag(
-                startLocationX: 12,
-                translation: CGSize(width: 96, height: 8)
-            )
-        )
-        #expect(
-            ReadingShellCompactNavigationState.shouldNavigateBackToSourcesOnDrag(
-                startLocationX: 64,
-                translation: CGSize(width: 96, height: 8)
-            ) == false
-        )
-        #expect(
-            ReadingShellCompactNavigationState.shouldNavigateBackToSourcesOnDrag(
-                startLocationX: 12,
-                translation: CGSize(width: 40, height: 8)
-            ) == false
-        )
-        #expect(
-            ReadingShellCompactNavigationState.shouldNavigateBackToSourcesOnDrag(
-                startLocationX: 12,
-                translation: CGSize(width: 96, height: 72)
-            ) == false
         )
     }
 
@@ -2183,89 +1904,6 @@ struct RSSReaderTests {
     }
 
     @Test
-    func readingShellSourceSwitchResetsArticleDetailSelectionAndTriggersReload() {
-        let appState = AppState()
-        let initialReloadID = appState.articleListReloadID
-        let feedID = UUID()
-        let articleID = UUID()
-
-        appState.selectReadingSource(.feed(feedID))
-        appState.selectedArticleID = articleID
-        appState.presentWebView(articleID: articleID, url: URL(string: "https://example.com/article")!)
-
-        let reloadIDBeforeSwitch = appState.articleListReloadID
-
-        appState.selectReadingSource(.inbox)
-
-        #expect(appState.selectedSidebarSelection == .inbox)
-        #expect(appState.selectedArticleID == nil)
-        #expect(appState.selectedDetailRoute == .none)
-        #expect(appState.presentedWebViewRoute == nil)
-        #expect(reloadIDBeforeSwitch != initialReloadID)
-        #expect(appState.articleListReloadID != reloadIDBeforeSwitch)
-    }
-
-    @Test
-    func readingShellSelectingSameSourceDoesNotResetSelectionOrTriggerReload() {
-        let appState = AppState()
-        let feedID = UUID()
-        let articleID = UUID()
-
-        appState.selectReadingSource(.feed(feedID))
-        appState.selectedArticleID = articleID
-
-        let reloadIDBeforeReselect = appState.articleListReloadID
-
-        appState.selectReadingSource(.feed(feedID))
-
-        #expect(appState.selectedSidebarSelection == .feed(feedID))
-        #expect(appState.selectedArticleID == articleID)
-        #expect(appState.selectedDetailRoute == .article(articleID))
-        #expect(appState.articleListReloadID == reloadIDBeforeReselect)
-    }
-
-    @Test
-    func readingShellSourcesFilterSwitchUpdatesActiveFilterWithoutBreakingNavigationContext() {
-        let appState = AppState()
-        let feedID = UUID()
-        let articleID = UUID()
-
-        appState.selectReadingSource(.feed(feedID))
-        appState.selectedArticleID = articleID
-        let reloadIDBeforeFilterSwitch = appState.articleListReloadID
-
-        appState.selectSourcesFilter(.starred)
-
-        #expect(appState.selectedSourcesFilter == .starred)
-        #expect(appState.selectedSidebarSelection == .feed(feedID))
-        #expect(appState.selectedArticleID == articleID)
-        #expect(appState.selectedDetailRoute == .article(articleID))
-        #expect(appState.presentedWebViewRoute == nil)
-        #expect(appState.articleListReloadID == reloadIDBeforeFilterSwitch)
-    }
-
-    @Test
-    func readingShellReapplyingSameSourcesFilterKeepsShellStateStable() {
-        let appState = AppState()
-        let articleID = UUID()
-        let webURL = URL(string: "https://example.com/sources-filter-article")!
-
-        appState.selectSourcesFilter(.unread)
-        appState.selectedArticleID = articleID
-        appState.presentWebView(articleID: articleID, url: webURL)
-
-        let reloadIDBeforeReapplyingFilter = appState.articleListReloadID
-
-        appState.selectSourcesFilter(.unread)
-
-        #expect(appState.selectedSourcesFilter == .unread)
-        #expect(appState.selectedArticleID == articleID)
-        #expect(appState.selectedDetailRoute == .webView(ArticleWebViewRoute(articleID: articleID, url: webURL)))
-        #expect(appState.presentedWebViewRoute == ArticleWebViewRoute(articleID: articleID, url: webURL))
-        #expect(appState.articleListReloadID == reloadIDBeforeReapplyingFilter)
-    }
-
-    @Test
     func folderSelectionInheritsActiveSourcesFilterForSelectedFolder() throws {
         let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
         let feeds = try harness.insertFeeds(
@@ -2602,36 +2240,6 @@ struct RSSReaderTests {
         let ungroupedFeeds = SidebarUngroupedFeeds.visibleFeeds(from: [groupedFeed])
 
         #expect(ungroupedFeeds.isEmpty)
-    }
-
-    @Test
-    func readingShellOpenArticleWebViewSetsPresentedRouteAndPreservesArticleContext() {
-        let appState = AppState()
-        let articleID = UUID()
-        let webURL = URL(string: "https://example.com/webview-article")!
-
-        appState.selectedArticleID = articleID
-        appState.presentWebView(articleID: articleID, url: webURL)
-
-        #expect(appState.selectedArticleID == articleID)
-        #expect(appState.selectedDetailRoute == .webView(ArticleWebViewRoute(articleID: articleID, url: webURL)))
-        #expect(appState.presentedWebViewRoute == ArticleWebViewRoute(articleID: articleID, url: webURL))
-    }
-
-    @Test
-    func readingShellClosingArticleWebViewRestoresArticleDetailRoute() {
-        let appState = AppState()
-        let articleID = UUID()
-        let webURL = URL(string: "https://example.com/webview-close")!
-
-        appState.selectedArticleID = articleID
-        appState.presentWebView(articleID: articleID, url: webURL)
-
-        appState.dismissPresentedWebView()
-
-        #expect(appState.selectedArticleID == articleID)
-        #expect(appState.selectedDetailRoute == .article(articleID))
-        #expect(appState.presentedWebViewRoute == nil)
     }
 
     @Test
